requirements.txt
# Sistema de Evaluación Automática con Rúbricas
# Requirements.txt - Versión optimizada

# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# GitHub API
PyGithub>=1.59.0
requests>=2.31.0

# OpenAI / GitHub Models
openai>=1.3.0

# Google Gemini (opcional)
google-generativeai>=0.3.0

# Data handling básico
pandas>=2.0.0
PyYAML>=6.0

# Visualization y reportes
plotly>=5.17.0
matplotlib>=3.7.0

# CLI utilities
rich>=13.6.0
tqdm>=4.66.0

# Export formats
openpyxl>=3.1.0  # Para Excel
fpdf2>=2.7.0     # Para PDF

# Testing (opcional)
pytest>=7.4.0
